refactor(agents): share marker/file-walk and manifest stamping helpers
Three sets of byte-identical duplication across the host installers:

- Hoist `skill_contents_have_tracedecay_marker` and the recursive
  `collect_regular_files` walker into `agents::mod` and route the Cursor
  (and, next commit, Codex) copies through them.
- Add `plugin_bundle::{stamp_manifest_version, set_mcp_command}` and route
  Claude, Cursor, and Codex manifest/MCP-command rewrites through them.
  Codex layers its scope-specific args/env on top of the shared base.

Behavior is byte-identical; the shared helpers reproduce each host's exact
output (pretty JSON + trailing newline, same marker clauses/walk order).
